Relocate exchange-rate and compact-formatting extensions to Bodu.Financial.Extensions
Move MoneyExchangeRateExtensions, MoneyOfTCurrencyExchangeRateExtensions, and
MoneyCompactFormattingExtensions out of Bodu.Financial into the
Bodu.Financial.Extensions namespace under src/Financial.Extensions/, organised
one-member-per-partial-file after the DateTimeExtensions pattern (the
compact-formatting private helpers stay in the class root file). This completes
the consistency follow-up so every Money and exchange-rate extension lives in
the dedicated extensions namespace alongside the sign/magnitude helpers.

Update the two ExchangeRate.cs crefs and the docs xrefs (the Financial apidoc
overview and the money guide) that referenced these classes by their former
Bodu.Financial UID.
