+++ Bodu.Financial/src/Financial.Extensions/MoneyExchangeRateExtensions.cs
@@ -1,0 +1,16 @@
+// ---------------------------------------------------------------------------------------------------------------
+// <copyright file="MoneyExchangeRateExtensions.cs" company="Bodu Pty. Ltd.">
+// Copyright (c) Bodu Pty. Ltd. All rights reserved.
+// </copyright>
+// ---------------------------------------------------------------------------------------------------------------
+
+namespace Bodu.Financial.Extensions;
+
+/// <summary>
+/// Provides extension methods that resolve a dated exchange rate from an <see cref="IDatedExchangeRateProvider" /> and
+/// apply it to a <see cref="Money" /> — the runtime-tagged counterpart of
+/// <see cref="MoneyOfTCurrencyExchangeRateExtensions" />.
+/// </summary>
+public static partial class MoneyExchangeRateExtensions
+{
+}
