+++ Bodu.Financial/src/Financial.Extensions/MoneyOfTCurrencyExchangeRateExtensions.cs
@@ -1,0 +1,15 @@
+// ---------------------------------------------------------------------------------------------------------------
+// <copyright file="MoneyOfTCurrencyExchangeRateExtensions.cs" company="Bodu Pty. Ltd.">
+// Copyright (c) Bodu Pty. Ltd. All rights reserved.
+// </copyright>
+// ---------------------------------------------------------------------------------------------------------------
+
+namespace Bodu.Financial.Extensions;
+
+/// <summary>
+/// Provides extension methods that resolve a dated exchange rate from an <see cref="IDatedExchangeRateProvider" /> and
+/// apply it to a <see cref="Money{TCurrency}" /> value.
+/// </summary>
+public static partial class MoneyOfTCurrencyExchangeRateExtensions
+{
+}
