+++ Bodu.Financial/src/Financial/ExchangeRate.cs
@@ -207,9 +207,9 @@ public readonly record struct ExchangeRate
     /// rate rather than multiplying by a pre-rounded reciprocal, avoiding a double-rounding step. Rounding is
     /// intentionally deferred to the money boundary so the rate object stays decoupled from the destination currency's
     /// minor-unit precision. Use
-    /// <see cref="MoneyOfTCurrencyExchangeRateExtensions.ConvertTo{TSource, TTarget}(Money{TSource}, IDatedExchangeRateProvider, DateOnly, ExchangeRateLookupOptions?, MidpointRounding)" />
+    /// <see cref="Bodu.Financial.Extensions.MoneyOfTCurrencyExchangeRateExtensions.ConvertTo{TSource, TTarget}(Money{TSource}, IDatedExchangeRateProvider, DateOnly, ExchangeRateLookupOptions?, MidpointRounding)" />
     /// ,
-    /// <see cref="MoneyExchangeRateExtensions.ConvertTo(Money, IDatedExchangeRateProvider, string, DateOnly, ExchangeRateLookupOptions?, MidpointRounding)" />
+    /// <see cref="Bodu.Financial.Extensions.MoneyExchangeRateExtensions.ConvertTo(Money, IDatedExchangeRateProvider, string, DateOnly, ExchangeRateLookupOptions?, MidpointRounding)" />
     /// , or <see cref="Money{TCurrency}.Convert{TTarget}(decimal, MidpointRounding)" /> to apply rounding at the
     /// destination precision.
     /// </remarks>
